requirements.txt
# DCL Ecommerce Requirements

# Django Framework
Django>=5.0,<5.1
django-environ>=0.11.2
django-extensions>=3.2.3
Pillow>=10.2.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL for production
dj-database-url>=2.1.0

# Authentication
django-allauth>=0.60.1

# Admin Enhancements
django-import-export>=3.3.6

# Forms
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2

# Celery (optional - for background tasks)
celery>=5.3.6
redis>=5.0.1

# Security
django-cors-headers>=4.3.1

# Development
django-debug-toolbar>=4.2.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# SSLCommerz
sslcommerz-lib>=1.0.7
requests>=2.31.0

# Testing
pytest>=8.0.0
pytest-django>=4.7.0
factory-boy>=3.3.0
